+++ CHANGELOG.md
@@ -56,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ### [0.2.1] - 2023-MM-DD
+- Added 'extraHeaders' param to Django services
 #### TODO
 
 -------------------------------------------------------